+++ business-logic/src/main/java/com/example/businesslogic/record/flower/UpdateFlowerPayload.java
@@ -6,8 +6,8 @@ public record UpdateFlowerPayload(
 
         @NotBlank(message = "Название не должно быть пустым")
         String title,
-        Double price,
-        Double extraCharge,
+        Integer price,
+        Integer extraCharge,
         Integer accountingQuantity,
         Integer actualQuantity
